requirements.txt
matplotlib
numpy
pandas
Pillow
pyyaml
pyperclip
selenium
timm
# --extra-index-url https://download.pytorch.org/whl/cu113
# torch == 1.12.1+cu113
# torchvision == 0.13.1+cu113
tqdm
webdriver_manager
